Update ssr.js
In your provided example you are using getServerProps, which re-fetches the new data on every page transition. This approach efectively renders the client-side caching useless and most of the Apollo optimisations are not in use. In the updated example I explain this situation and proide an alternative that takes advantage of client side caching on page transitions.